+++ packages/vuelidate/index.d.ts
@@ -40,7 +40,7 @@ export interface ValidatorResponse {
   [key: string]: any
 }
 
-export type ValidatorFn <T = any, K = object, S = any> = (value: T, siblingState: K, vm: Vue & S) => boolean | ValidatorResponse | Promise<boolean | ValidatorResponse>;
+export type ValidatorFn <T = any, K = object, S = any> = (value: T, siblingState: K, vm: S) => boolean | ValidatorResponse | Promise<boolean | ValidatorResponse>;
 
 export interface ValidationRuleWithoutParams <T = any> {
   $validator: ValidatorFn<T>
@@ -112,6 +112,7 @@ type BaseValidation <
   // const validationMethods
   readonly $touch: () => void
   readonly $reset: () => void
+  readonly $commit: () => void
   readonly $validate: () => Promise<boolean>
 };
 
@@ -170,7 +171,8 @@ export interface GlobalConfig {
   $stopPropagation?: boolean
   $autoDirty?: boolean
   $lazy?: boolean,
-  $externalResults?: ServerErrors | Ref<ServerErrors> | UnwrapRef<ServerErrors>
+  $externalResults?: ServerErrors | Ref<ServerErrors> | UnwrapRef<ServerErrors>,
+  $rewardEarly?: boolean
 }
 
 export function useVuelidate(globalConfig?: GlobalConfig): Ref<Validation>;