+++ packages/vuelidate/src/index.js
@@ -77,6 +77,7 @@ function nestedValidations ({ $scope }) {
  * @property {Ref<Object>} [$externalResults] - External error messages, like from server validation.
  * @property {Boolean} [$autoDirty] - Should the form watch for state changed, and automatically set `$dirty` to true.
  * @property {Boolean} [$lazy] - Should the validations be lazy, and run only after they are dirty
+ * @property {Boolean} [$rewardEarly] - Once valid, re-runs property validators only on manual calls of $commit
  */
 
 /**