+++ packages/components/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.10](https://github.com/vuelidate/vuelidate/compare/@vuelidate/components@1.1.9...@vuelidate/components@1.1.10) (2021-10-22)
+
+**Note:** Version bump only for package @vuelidate/components
+
+
+
+
+
+## [1.1.9](https://github.com/vuelidate/vuelidate/compare/@vuelidate/components@1.1.8...@vuelidate/components@1.1.9) (2021-10-20)
+
+**Note:** Version bump only for package @vuelidate/components
+
+
+
+
+
+## [1.1.8](https://github.com/vuelidate/vuelidate/compare/@vuelidate/components@1.1.7...@vuelidate/components@1.1.8) (2021-10-15)
+
+**Note:** Version bump only for package @vuelidate/components
+
+
+
+
+
+## [1.1.7](https://github.com/vuelidate/vuelidate/compare/@vuelidate/components@1.1.6...@vuelidate/components@1.1.7) (2021-10-06)
+
+**Note:** Version bump only for package @vuelidate/components
+
+
+
+
+
 ## [1.1.6](https://github.com/vuelidate/vuelidate/compare/@vuelidate/components@1.1.5...@vuelidate/components@1.1.6) (2021-09-17)
 
 **Note:** Version bump only for package @vuelidate/components